Set up NuGet, added Entity Framework reference.
The Phlebotomist.Model project uses Entity Framework 5.0.0, so that
needed to be installed from the Package Manager Console in Visual Studio
using the following command.

PM> install-package entityframework -version 5.0.0.0 -project BrigadeRouletteConsole

I am not sure if indicating the project (with the -project flag) is
necessary or correct, as the output from the command seemed to indicate
that it tried to install Entity Framework for all projects in the
solution.